Reorganize into a standard uv src layout
- pyproject.toml + uv.lock at the repo root, with a hatchling build
  backend and a `parlor` console script: quick start is now
  `uv sync && uv run parlor`, no more `cd src`
- Python modules move to the src/parlor/ package; the frontend
  (index.html + static/) to src/parlor/web/
- whisper_features.py merged into turn_detector.py — it was a private
  implementation detail of its only consumer (vendored-code attribution
  preserved)
- tests/ and benchmarks/ move to the repo root; every sys.path hack is
  replaced with `from parlor ...` package imports (the editable install
  makes them resolvable), and tests reach benchmarks/fixtures.py via
  pytest's pythonpath setting
- conftest spawns the suite's server as `python -m parlor.server`

Verified: uv sync + full e2e suite (86 passed, 1 skipped).
