+++ pyproject.toml
@@ -18,9 +18,17 @@ dependencies = [
     "onnxruntime>=1.24.4",
 ]
 
+[project.scripts]
+parlor = "parlor.server:main"
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
 [dependency-groups]
 dev = ["pytest>=8.4"]
 
 [tool.pytest.ini_options]
 testpaths = ["tests"]
+pythonpath = ["benchmarks"]  # tests synthesize speech via benchmarks/fixtures.py
