+++ src/parlor/__init__.py
@@ -1,0 +1,1 @@
+"""Parlor — on-device, real-time multimodal AI (voice + vision)."""
