+++ tests/test_transcription_stability.py
