+++ tests/test_translation.py
@@ -3,12 +3,7 @@
 the spoken command and the UI escape hatch (set_mode).
 """
 
-import sys
-from pathlib import Path
-
-sys.path.insert(0, str(Path(__file__).resolve().parent.parent))
-
-import util  # noqa: E402
+import util
 
 
 def switch_by_voice(session, fixture, target, tries=2):
